+++ Assets/Scripts/Enemies/Boss/EnemyMiddleBoss5a.cs
@@ -55,7 +55,7 @@ public class EnemyMiddleBoss5a : EnemyUnit, IEnemyBossMain
         int frame = APPEARANCE_TIME * Application.targetFrameRate / 1000;
 
         for (int i = 0; i < frame; ++i) {
-            float t_pos_y = AC_Ease.ac_ease[EaseType.OutQuad].Evaluate((float) (i+1) / frame);
+            float t_pos_y = AC_Ease.ac_ease[(int)EaseType.OutQuad].Evaluate((float) (i+1) / frame);
             
             float position_y = Mathf.Lerp(init_position_y, TARGET_POSITION.y, t_pos_y);
             transform.position = new Vector3(transform.position.x, position_y, transform.position.z);
@@ -82,7 +82,7 @@ public class EnemyMiddleBoss5a : EnemyUnit, IEnemyBossMain
         float init_position_y = transform.position.y;
 
         for (int i = 0; i < frame; ++i) {
-            float t_pos_y = AC_Ease.ac_ease[EaseType.InQuad].Evaluate((float) (i+1) / frame);
+            float t_pos_y = AC_Ease.ac_ease[(int)EaseType.InQuad].Evaluate((float) (i+1) / frame);
             
             float position_y = Mathf.Lerp(init_position_y, Size.GAME_BOUNDARY_BOTTOM - 8f, t_pos_y);
             transform.position = new Vector3(transform.position.x, position_y, transform.position.z);
